// src/app/interfaces/registration.ts
export interface UserRegistration {
    firstName: string;
    lastName: string;
    email: string;
    phone: number;
    state: string;
    lga: string;
    homeAddress: string;
    businessAddress: string;
    gender: string;
    loanCategory: string;
    pin: string;
    status: string;
    loanStatus: string;
    password: number;
    passwordString: number;
}